#46 fix: 반복 내역 삭제 성공 시, AddView 닫기

+++ Floney/Source/Main/Home/SettingBook/ViewModel/ManageRepeatLineViewModel.swift
@@ -9,15 +9,13 @@ import Foundation
 import Combine
 
 class ManageRepeatLineViewModel : ObservableObject {
-#if DEBUG
-    let projectMode = "dev"
-#else
-    let projectMode = "prod"
-#endif
+    var alertManager = AlertManager.shared
     @Published var tokenViewModel = TokenReissueViewModel()
     @Published var categoryType = ""
     @Published var repeatLineList : [RepeatLineResponse] = []
     
+    @Published var isApiCalling: Bool = false
+    @Published var successStatus: Bool = false
     private var cancellableSet: Set<AnyCancellable> = []
     var dataManager: ManageRepeatLineProtocol
     
@@ -47,14 +45,20 @@ class ManageRepeatLineViewModel : ObservableObject {
     }
     
     func deleteRepeatLine(repeatLineId: Int) {
+        guard !isApiCalling else { return }
+        isApiCalling = true
         let request = DeleteRepeatLineRequest(repeatLineId: repeatLineId)
         dataManager.deleteRepeatLine(parameters: request)
             .sink { completion in
                 switch completion {
                 case .finished:
+                    self.isApiCalling = false
+                    self.successStatus = true
+                    self.alertManager.update(showAlert: true, message: "삭제가 완료되었습니다.", buttonType: .green)
                     print("반복 내역 삭제 성공")
                 case .failure(let error):
                     print(error)
+                    self.isApiCalling = false
                     self.createAlert(with: error, retryRequest: {
                         self.deleteRepeatLine(repeatLineId: repeatLineId)
                     })
@@ -66,14 +70,20 @@ class ManageRepeatLineViewModel : ObservableObject {
     }
     
     func deleteAllRepeatLine(bookLineKey: Int) {
+        guard !isApiCalling else { return }
+        isApiCalling = true
         let request = DeleteAllRepeatLineRequest(bookLineKey: bookLineKey)
         dataManager.deleteAllRepeatLine(parameters: request)
             .sink { completion in
                 switch completion {
                 case .finished:
+                    self.isApiCalling = false
+                    self.successStatus = true
+                    self.alertManager.update(showAlert: true, message: "삭제가 완료되었습니다.", buttonType: .green)
                     print("반복 내역 모두 삭제 성공")
                 case .failure(let error):
                     print(error)
+                    self.isApiCalling = false
                     self.createAlert(with: error, retryRequest: {
                         self.deleteAllRepeatLine(bookLineKey: bookLineKey)
                     })
